+++ src/termux_tasker/ui/screens/install_runner.py
@@ -7,6 +7,7 @@ from textual.widgets import Button
 
 from termux_tasker.config import RunnerMetadata
 from termux_tasker.ui.base import ButtonConfig, ButtonLayout, MenuScreen
+from termux_tasker.ui.screens._ui_utils import go_home
 from termux_tasker.ui.screens._utils import termux_app
 
 
@@ -18,6 +19,7 @@ class InstallRunnerScreen(MenuScreen):
         super().__init__(
             menu_items=[ButtonConfig("install", "Install", variant="primary", layout=ButtonLayout.BOTTOM)],
             description=meta.general.description or "",
+            show_home_button=True,
             show_back_button=True,
         )
         self.title = "Install Runner"
@@ -30,3 +32,8 @@ class InstallRunnerScreen(MenuScreen):
         termux_app(self).push_screen(
             InstallRunnerVersionScreen(self.tmp_runner_folder)
         )
+
+    @on(Button.Pressed, "#home")
+    def on_home_button_pressed(self, event: Button.Pressed) -> None:
+        event.stop()
+        go_home(self)