+++ src/termux_tasker/ui/screens/install_task_version.py
@@ -22,6 +22,7 @@ from termux_tasker.ui.base import (
     ConfirmationScreen,
     InputScreen,
 )
+from termux_tasker.ui.screens._ui_utils import go_home
 from termux_tasker.ui.screens._utils import (
     termux_app,
     fetch_git_tags,
@@ -43,7 +44,7 @@ class InstallTaskVersionScreen(MenuScreen):
         self._task_meta = meta
         self._id_to_tag: dict[str, str] = {}
 
-        super().__init__([], show_back_button=True)
+        super().__init__([], show_home_button=True, show_back_button=True)
         self.title = "Task Version"
         self.sub_title = meta.general.name
         self._loaded = False
@@ -303,3 +304,8 @@ class InstallTaskVersionScreen(MenuScreen):
             termux_app(self).pop_screen()
 
         termux_app(self).push_screen(TaskMenuScreen(target_dir))
+
+    @on(Button.Pressed, "#home")
+    def on_home_button_pressed(self, event: Button.Pressed) -> None:
+        event.stop()
+        go_home(self)