+++ src/termux_tasker/ui/screens/task_menu.py
@@ -24,7 +24,7 @@ from termux_tasker.ui.screens._utils import (
     termux_app,
     copy_to_tmp,
 )
-from termux_tasker.ui.screens._ui_utils import ask_validated_input
+from termux_tasker.ui.screens._ui_utils import ask_validated_input, go_home
 from termux_tasker.ui.screens.properties import PropertiesScreen
 from termux_tasker.ui.screens.widgets.description import (
     KeyValueEntry,
@@ -60,7 +60,7 @@ class TaskMenuScreen(MenuScreen):
         )
         items = self._build_items(meta, settings)
 
-        super().__init__(items, description_widget=self._state, show_back_button=True)
+        super().__init__(items, description_widget=self._state, show_home_button=True, show_back_button=True)
         self.title = "Task"
         self.sub_title = meta.general.name
         self._poll_timer: Any = None
@@ -254,3 +254,8 @@ class TaskMenuScreen(MenuScreen):
 
         shutil.rmtree(self.task_path, ignore_errors=True)
         termux_app(self).pop_screen()   # noqa
+
+    @on(Button.Pressed, "#home")
+    def on_home_button_pressed(self, event: Button.Pressed) -> None:
+        event.stop()
+        go_home(self)